+++ SparkyFitnessMCP/src/services/reportService.ts
@@ -61,3 +61,62 @@ export async function getWeeklyReport(userId: string, endDate?: string): Promise
 
   return report;
 }
+
+type McpReportDateQuery = { date?: string; start_date?: string; end_date?: string };
+
+function mcpReportDateRange(query: McpReportDateQuery = {}): { startDate: string; endDate: string } {
+  const today = new Date().toISOString().slice(0, 10);
+  const date = query.date || undefined;
+  const startDate = date || query.start_date || today;
+  const endDate = date || query.end_date || startDate;
+  return { startDate, endDate };
+}
+
+export async function getDailyReport(userId: string, params: McpReportDateQuery = {}): Promise<Record<string, unknown>> {
+  const { startDate, endDate } = mcpReportDateRange(params);
+
+  return withClient(userId, async (client) => {
+    const nutrition = await client.query(
+      `SELECT entry_date,
+              SUM(COALESCE(calories, 0)) AS calories,
+              SUM(COALESCE(protein, 0)) AS protein,
+              SUM(COALESCE(carbs, 0)) AS carbs,
+              SUM(COALESCE(fat, 0)) AS fat,
+              SUM(COALESCE(dietary_fiber, 0)) AS fiber
+       FROM food_entries
+       WHERE entry_date BETWEEN $1 AND $2
+       GROUP BY entry_date
+       ORDER BY entry_date ASC`,
+      [startDate, endDate],
+    );
+
+    const exercise = await client.query(
+      `SELECT entry_date,
+              SUM(COALESCE(calories_burned, 0)) AS exercise_calories,
+              SUM(COALESCE(duration_minutes, 0)) AS exercise_minutes,
+              SUM(COALESCE(steps, 0)) AS steps
+       FROM exercise_entries
+       WHERE entry_date BETWEEN $1 AND $2
+       GROUP BY entry_date
+       ORDER BY entry_date ASC`,
+      [startDate, endDate],
+    ).catch(() => ({ rows: [] as Record<string, unknown>[] }));
+
+    const water = await client.query(
+      `SELECT entry_date, SUM(COALESCE(amount_ml, 0)) AS water_ml
+       FROM water_entries
+       WHERE entry_date BETWEEN $1 AND $2
+       GROUP BY entry_date
+       ORDER BY entry_date ASC`,
+      [startDate, endDate],
+    ).catch(() => ({ rows: [] as Record<string, unknown>[] }));
+
+    return {
+      start_date: startDate,
+      end_date: endDate,
+      nutrition: nutrition.rows,
+      exercise: exercise.rows,
+      water: water.rows,
+    };
+  });
+}
