+++ SparkyFitnessMCP/src/tools/register.ts
@@ -11,8 +11,6 @@ import { registerProfileTools } from "./profile.js";
 import { registerHabitTools } from "./habits.js";
 import { registerWizardTools } from "./wizard.js";
 import { registerReportTools } from "./report.js";
-import { registerRawTools } from "./raw.js";
-
 /**
  * Registers all MCP tools for the authenticated user.
  * Each tool module is imported and its register function called.
@@ -31,5 +29,4 @@ export function registerAllTools(server: McpServer, userId: string): void {
   registerHabitTools(server, userId);
   registerWizardTools(server, userId);
   registerReportTools(server, userId);
-  registerRawTools(server, userId);
 }
